requirements.txt
kafka-python
influxdb-client
pandas
matplotlib
json5
datetime